+++ parallel-consumer-core/src/test-integration/java/io/confluent/parallelconsumer/integrationTests/MultiInstanceHighVolumeTest.java
@@ -8,7 +8,7 @@ import io.confluent.csid.utils.ProgressBarUtils;
 import io.confluent.csid.utils.StringUtils;
 import io.confluent.csid.utils.TrimListRepresentation;
 import io.confluent.parallelconsumer.ParallelConsumerOptions;
-import io.confluent.parallelconsumer.ParentParallelEoSStreamProcessor;
+import io.confluent.parallelconsumer.ParallelEoSStreamProcessor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import me.tongfei.progressbar.ProgressBar;
@@ -64,9 +64,9 @@ class MultiInstanceHighVolumeTest extends BrokerIntegrationTest<String, String> 
         produceMessages(inputTopicName, expectedKeys, expectedMessageCount);
 
         // setup
-        ParentParallelEoSStreamProcessor<String, String> pcOne = buildPc(inputTopicName, expectedMessageCount);
-        ParentParallelEoSStreamProcessor<String, String> pcTwo = buildPc(inputTopicName, expectedMessageCount);
-        ParentParallelEoSStreamProcessor<String, String> pcThree = buildPc(inputTopicName, expectedMessageCount);
+        ParallelEoSStreamProcessor<String, String> pcOne = buildPc(inputTopicName, expectedMessageCount);
+        ParallelEoSStreamProcessor<String, String> pcTwo = buildPc(inputTopicName, expectedMessageCount);
+        ParallelEoSStreamProcessor<String, String> pcThree = buildPc(inputTopicName, expectedMessageCount);
 
         // run
         var consumedByOne = Collections.synchronizedList(new ArrayList<ConsumerRecord<?, ?>>());
@@ -112,7 +112,7 @@ class MultiInstanceHighVolumeTest extends BrokerIntegrationTest<String, String> 
 
     Integer barId = 0;
 
-    private ProgressBar run(final int expectedMessageCount, final ParentParallelEoSStreamProcessor<String, String> pc, List<ConsumerRecord<?, ?>> consumed) {
+    private ProgressBar run(final int expectedMessageCount, final ParallelEoSStreamProcessor<String, String> pc, List<ConsumerRecord<?, ?>> consumed) {
         ProgressBar bar = ProgressBarUtils.getNewMessagesBar(log, expectedMessageCount);
         bar.setExtraMessage("#" + barId);
         pc.setMyId(Optional.of("id: " + barId));
@@ -147,8 +147,8 @@ class MultiInstanceHighVolumeTest extends BrokerIntegrationTest<String, String> 
 //        return new ProducerRecord<>(outputName, record.key(), "data");
     }
 
-    private ParentParallelEoSStreamProcessor<String, String> buildPc(final String inputName,
-                                                                     final int expectedMessageCount) {
+    private ParallelEoSStreamProcessor<String, String> buildPc(final String inputName,
+                                                               final int expectedMessageCount) {
         log.debug("Starting test");
 //        KafkaProducer<String, String> newProducer = kcu.createNewProducer(commitMode.equals(PERIODIC_TRANSACTIONAL_PRODUCER));
 
@@ -156,7 +156,7 @@ class MultiInstanceHighVolumeTest extends BrokerIntegrationTest<String, String> 
         consumerProps.put(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, maxPoll);
         KafkaConsumer<String, String> newConsumer = kcu.createNewConsumer(false, consumerProps);
 
-        var pc = new ParentParallelEoSStreamProcessor<String, String>(ParallelConsumerOptions.<String, String>builder()
+        var pc = new ParallelEoSStreamProcessor<String, String>(ParallelConsumerOptions.<String, String>builder()
                 .ordering(order)
                 .consumer(newConsumer)
                 .commitMode(commitMode)
